first

+++ lib/python/Screens/Standby.py
@@ -1,5 +1,5 @@
 # -*- coding: utf-8 -*-
-import os
+from os.path import isfile
 import struct
 import RecordTimer
 import Components.ParentalControl
@@ -124,7 +124,7 @@ class StandbyScreen(Screen):
 		print("[Standby] enter standby")
 		BoxInfo.setItem("StandbyState", True)
 
-		if os.path.exists("/usr/script/standby_enter.sh"):
+		if isfile("/usr/script/standby_enter.sh"):
 			Console().ePopen("/usr/script/standby_enter.sh")
 
 		self["actions"] = ActionMap(["StandbyActions"],
@@ -181,11 +181,12 @@ class StandbyScreen(Screen):
 		else:
 			self.avswitch.setInput("AUX")
 
-		try:
-			print("[Standby] Write to /proc/stb/hdmi/output")
-			open("/proc/stb/hdmi/output", "w").write("off")
-		except:
-			print("[Standby] Write to /proc/stb/hdmi/output failed.")
+		if isfile("/proc/stb/hdmi/output"):
+			try:
+				print("[Standby] Write to /proc/stb/hdmi/output")
+				open("/proc/stb/hdmi/output", "w").write("off")
+			except:
+				print("[Standby] Write to /proc/stb/hdmi/output failed.")
 
 		if BoxInfo.getItem("AmlogicFamily"):
 			try:
@@ -237,7 +238,7 @@ class StandbyScreen(Screen):
 			RecordTimer.RecordTimerEntry.stopTryQuitMainloop()
 		self.avswitch.setInput("ENCODER")
 		self.leaveMute()
-		if os.path.exists("/usr/script/standby_leave.sh"):
+		if isfile("/usr/script/standby_leave.sh"):
 			Console().ePopen("/usr/script/standby_leave.sh")
 		if config.usage.remote_fallback_import_standby.value and not config.clientmode.enabled.value:
 			ImportChannels()
@@ -254,14 +255,15 @@ class StandbyScreen(Screen):
 		BoxInfo.setItem("StandbyState", False)
 		self.close(True)
 
-		if os.path.exists("/usr/script/StandbyLeave.sh"):
+		if isfile("/usr/script/StandbyLeave.sh"):
 			Console().ePopen("/usr/script/StandbyLeave.sh")
 
-		try:
-			print("[Standby] Write to /proc/stb/hdmi/output")
-			open("/proc/stb/hdmi/output", "w").write("on")
-		except:
-			print("[Standby] Write to /proc/stb/hdmi/output failed.")
+		if isfile("/proc/stb/hdmi/output"):
+			try:
+				print("[Standby] Write to /proc/stb/hdmi/output")
+				open("/proc/stb/hdmi/output", "w").write("on")
+			except:
+				print("[Standby] Write to /proc/stb/hdmi/output failed.")
 
 		if BoxInfo.getItem("AmlogicFamily"):
 			try:
@@ -456,7 +458,7 @@ class TryQuitMainloop(MessageBox):
 			if self.retval == QUIT_SHUTDOWN:
 				config.misc.DeepStandby.value = True
 				if not inStandby:
-					if os.path.exists("/usr/script/standby_enter.sh"):
+					if isfile("/usr/script/standby_enter.sh"):
 						Console().ePopen("/usr/script/standby_enter.sh")
 			elif not inStandby:
 				config.misc.RestartUI.value = True
